<fix/likes>: fixed response

+++ src/main/java/com/api/controller/websocket/likes/LikesController.java
@@ -3,6 +3,8 @@ package com.api.controller.websocket.likes;
 import java.security.Principal;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
@@ -11,7 +13,9 @@ import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Controller;
 
+import com.api.model.ContentPosting;
 import com.api.model.Likes;
+import com.api.repository.ContentPostingRepository;
 import com.api.repository.LikesRepository;
 import com.api.security.StompPrincipal;
 
@@ -21,6 +25,8 @@ public class LikesController {
   private LikesRepository likesRepository;
   @Autowired
   private SimpMessagingTemplate messagingTemplate;
+  @Autowired
+  ContentPostingRepository contentPostingRepository;
 
   @MessageMapping("/like/{contentId}")
   public void toggleLike(
@@ -30,14 +36,22 @@ public class LikesController {
     if (principal == null || principal.getName() == null) {
       throw new SecurityException("Access is denied for: " + contentId);
     }
+    Optional<ContentPosting> contentPostingOpt = contentPostingRepository.findById(contentId);
+    if (!contentPostingOpt.isPresent()) {
+      throw new SecurityException("Not found for: " + contentId);
+    }
+    ContentPosting contentPosting = contentPostingOpt.get();
     if (principal instanceof StompPrincipal stompPrincipal) {
       if (likesRepository.existsByContentIdAndUserId(contentId, stompPrincipal.getUserId())) {
         likesRepository.deleteByUserIdAndContentId(stompPrincipal.getUserId(), contentId);
+        contentPosting.setLikeCount(contentPosting.getLikeCount() - 1);
       } else {
         likes.setCreatedAt(ZonedDateTime.now(ZoneId.of("Asia/Ho_Chi_Minh")));
         likesRepository.save(likes);
+        contentPosting.setLikeCount(contentPosting.getLikeCount() + 1);
       }
-      messagingTemplate.convertAndSend("/topic/contents/" + contentId, likes);
+      contentPostingRepository.save(contentPosting);
+      messagingTemplate.convertAndSend("/topic/contents/" + contentId, Map.of("likes", contentPosting.getLikeCount()));
     } else {
       throw new SecurityException("Invalid principal type");
     }
